requirements.txt
django
djangorestframework
django-ninja
rav
pillow